+++ CHANGELOG.md
@@ -4,11 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.2] - 2019-03-01
+
+### Added
+
+* **tgssub** action
+    * Substitutes in alternate sname (cifs) or SPN (ldap/computer.domain.com) into an existing service ticket
+
 
 ## [1.4.1] - 2019-02-25
 
 ### Added
-* **asktgs** action
+* to **asktgs** action
     * /enctype:[RC4/AES128/AES256/DES] now forces that particular encryption type in the TGS-REQ
 
 ### Changed